UINT, ULONG, DWORD, U32... Aaargh! They are irritating! So I replaced all those types to (u)intN_t. (Of course, true Windows types are not replaced)
I think some of uint32_t should be replaced by uint_least32_t, unsigned int/long, size_t, off_t or anything else.

+++ src/Loop.h
@@ -14,9 +14,9 @@ struct Loop							//Sample Block
 	{ }
 
 	int loopStatus;
-	ULONG loopType;				//can't think of it off my head, look it up in DLS specs.  I set it to 0 always
-	BYTE loopStartMeasure;
-	BYTE loopLengthMeasure;
-	ULONG loopStart;
-	ULONG loopLength;
+	uint32_t loopType;				//can't think of it off my head, look it up in DLS specs.  I set it to 0 always
+	uint8_t loopStartMeasure;
+	uint8_t loopLengthMeasure;
+	uint32_t loopStart;
+	uint32_t loopLength;
 };
