+++ src/VGMFile.cpp
@@ -9,7 +9,7 @@ DECLARE_MENU(VGMFile)
 
 using namespace std;
 
-VGMFile::VGMFile(FileType fileType, const string& fmt, RawFile* theRawFile, ULONG offset, ULONG length, wstring theName)
+VGMFile::VGMFile(FileType fileType, const string& fmt, RawFile* theRawFile, uint32_t offset, uint32_t length, wstring theName)
 : VGMContainerItem(this, offset, length),
   rawfile(theRawFile),
   bUsingRawFile(true),
@@ -30,22 +30,22 @@ VGMFile::~VGMFile(void)
 // this as an item because we do not want the VGMFile to be itself an item in the Item View
 void VGMFile::AddToUI(VGMItem* parent, VOID* UI_specific)
 {
-	for (UINT i=0; i<containers.size(); i++)
+	for (uint32_t i=0; i<containers.size(); i++)
 	{
-		for (UINT j=0; j<containers[i]->size(); j++)
+		for (uint32_t j=0; j<containers[i]->size(); j++)
 			(*containers[i])[j]->AddToUI(this, UI_specific);
 	}
 }
 
 // All we do is remove "return this", because we do not want to select the entire file when we click
 // an unspecified region of the file.
-VGMItem* VGMFile::GetItemFromOffset(ULONG offset)
+VGMItem* VGMFile::GetItemFromOffset(uint32_t offset)
 {
 	if (IsItemAtOffset(offset))				//if the offset is within this item
 	{
-		for (UINT i=0; i<containers.size(); i++)
+		for (uint32_t i=0; i<containers.size(); i++)
 		{
-			for (UINT j=0; j<containers[i]->size(); j++)	
+			for (uint32_t j=0; j<containers[i]->size(); j++)	
 			{
 				VGMItem* foundItem = (*containers[i])[j]->GetItemFromOffset(offset);
 				if (foundItem)
@@ -74,7 +74,7 @@ bool VGMFile::OnSaveAsRaw()
 	if (filepath.length() != 0)
 	{
 		bool result;
-		BYTE* buf = new BYTE[unLength];		//create a buffer the size of the file
+		uint8_t* buf = new uint8_t[unLength];		//create a buffer the size of the file
 		GetBytes(dwOffset, unLength, buf);
 		result = pRoot->UI_WriteBufferToFile(filepath.c_str(), buf, unLength);
 		delete[] buf;
@@ -176,12 +176,12 @@ void VGMFile::UseRawFileData()
 
 
 
-UINT VGMFile::GetBytes(UINT nIndex, UINT nCount, void* pBuffer)
+uint32_t VGMFile::GetBytes(uint32_t nIndex, uint32_t nCount, void* pBuffer)
 {
 	// if unLength != 0, verify that we're within the bounds of the file, and truncate num read bytes to end of file
 	if (unLength != 0)
 	{
-		UINT endOff = dwOffset+unLength;
+		uint32_t endOff = dwOffset+unLength;
 		assert (nIndex >= dwOffset && nIndex < endOff);
 		if (nIndex + nCount > endOff)
 			nCount = endOff - nIndex; 
@@ -206,7 +206,7 @@ UINT VGMFile::GetBytes(UINT nIndex, UINT nCount, void* pBuffer)
 // VGMHeader
 // *********
 
-VGMHeader::VGMHeader(VGMItem* parItem, ULONG offset, ULONG length, const wchar_t* name)
+VGMHeader::VGMHeader(VGMItem* parItem, uint32_t offset, uint32_t length, const wchar_t* name)
 : VGMContainerItem(parItem->vgmfile, offset, length, name)
 {
 	//AddContainer<VGMItem>(items);
@@ -217,17 +217,17 @@ VGMHeader::~VGMHeader()
 	//DeleteVect<VGMItem>(items);
 }
 
-void VGMHeader::AddPointer(ULONG offset, ULONG length, ULONG destAddress, bool notNull, const wchar_t *name)
+void VGMHeader::AddPointer(uint32_t offset, uint32_t length, uint32_t destAddress, bool notNull, const wchar_t *name)
 {
 	localitems.push_back(new VGMHeaderItem(this, VGMHeaderItem::HIT_POINTER, offset, length, name));
 }
 
-void VGMHeader::AddTempo(ULONG offset, ULONG length, const wchar_t *name)
+void VGMHeader::AddTempo(uint32_t offset, uint32_t length, const wchar_t *name)
 {
 	localitems.push_back(new VGMHeaderItem(this, VGMHeaderItem::HIT_TEMPO, offset, length, name));
 }
 
-void VGMHeader::AddSig(ULONG offset, ULONG length, const wchar_t *name)
+void VGMHeader::AddSig(uint32_t offset, uint32_t length, const wchar_t *name)
 {
 	localitems.push_back(new VGMHeaderItem(this, VGMHeaderItem::HIT_SIG, offset, length, name));
 }
@@ -236,7 +236,7 @@ void VGMHeader::AddSig(ULONG offset, ULONG length, const wchar_t *name)
 // VGMHeaderItem
 // *************
 
-VGMHeaderItem::VGMHeaderItem(VGMHeader* hdr, HdrItemType theType, ULONG offset, ULONG length, const wchar_t* name)
+VGMHeaderItem::VGMHeaderItem(VGMHeader* hdr, HdrItemType theType, uint32_t offset, uint32_t length, const wchar_t* name)
 : VGMItem(hdr->vgmfile, offset, length, name, CLR_HEADER), type(theType)
 {
 }
